instantiate ACE_Atomic_Op with int instead of u_long to match what's in libACE

+++ examples/Threads/future1.cpp
@@ -29,7 +29,7 @@
 
 #if defined (ACE_HAS_THREADS)
 
-typedef ACE_Atomic_Op<ACE_Thread_Mutex, u_long> ATOMIC_INT;
+typedef ACE_Atomic_Op<ACE_Thread_Mutex, int> ATOMIC_INT;
 
 // a counter for the tasks..
 static ATOMIC_INT task_count (0);
@@ -399,7 +399,7 @@ main (int, char *[])
 }
 
 #if defined (ACE_TEMPLATES_REQUIRE_SPECIALIZATION)
-template class ACE_Atomic_Op<ACE_Thread_Mutex, u_long>;
+template class ACE_Atomic_Op<ACE_Thread_Mutex, int>;
 #endif /* ACE_TEMPLATES_REQUIRE_SPECIALIZATION */
 
 #else
